Use minimal docker image for runtime
We use the python:latest-dev image to build the dependencies and install the libraries (it includes apk and uv by default).
Then we copy the installed libraries to an image based on python:latest which does not include any package-managers or shells and thus has a smaller attack surface.

This way we also no longer have to deal with version updates for the packages apk installs (as its now only one package and no longer includes the python version).

RISDEV-0000